requirements.txt
boto3==1.36.24
python-dotenv==1.0.1
pymysql==1.1.1
psycopg2-binary==2.9.10
jinja2==3.1.5
pytest==8.3.4
cryptography==44.0.0
